Remove sudo from build_and_check_circle_config
Interested to get your thoughts. I have a docker image that doesn't have sudo installed. So I can't rely on sudo. I wonder why this was setup in the first place, but it was presumably because the file permissions of your image were more sensibly locked down.

For now I will install sudo on the CI image to get this to pass. Just curious to hear your thoughts.